replaced ezid_ui_tests.py with the docker version
